Allow relationship methods to be defined on the transformer
A method may be defined for each include (providing it has the
same name as the include/relation). In Laravel, this method will
receive the query builder for that relationship. The result of the
method will be used as the related resource rather than getting it
as a property off the base resource.